review fixes + stabilize tests
