--- frontend/kesaseteli/employer/src/backend-api/backend-endpoints.ts
@@ -1,9 +1,0 @@
-import BackendPath from 'kesaseteli/employer/backend-api/backend-path';
-
-const BackendEndpoint: Record<string, BackendPath> = {
-  COMPANY: '/v1/company/',
-  USER: '/oidc/userinfo/',
-  LOGIN: '/oidc/authenticate/',
-  LOGOUT: '/oidc/logout/',
-} as const;
-export default BackendEndpoint;
