--- frontend/kesaseteli/employer/src/backend-api/backend-path.d.ts
@@ -1,7 +1,0 @@
-type BackendPath =
-  | '/v1/company/'
-  | '/oidc/userinfo/'
-  | '/oidc/authenticate/'
-  | '/oidc/logout/';
-
-export default BackendPath;
