--- frontend/kesaseteli/employer/src/backend-api/get-backend-domain.ts
@@ -1,4 +1,0 @@
-const getBackendDomain = (): string =>
-  process.env.NEXT_PUBLIC_BACKEND_URL || 'https://localhost:8000';
-
-export default getBackendDomain;
