--- frontend/kesaseteli/employer/src/backend-api/get-backend-url.ts
@@ -1,7 +1,0 @@
-import BackendPath from 'kesaseteli/employer/backend-api/backend-path';
-import getBackendDomain from 'kesaseteli/employer/backend-api/get-backend-domain';
-
-const getBackendUrl = (path: BackendPath): string =>
-  `${getBackendDomain()}${path}`;
-
-export default getBackendUrl;
