+++ frontend/kesaseteli/employer/src/hooks/useCompanyQuery.ts
@@ -1,4 +1,4 @@
-import backendEndpoint from 'kesaseteli/employer/backend-api/backend-endpoints';
+import { BackendEndpoint } from 'kesaseteli/employer/backend-api/backend-api';
 import handleResponse from 'kesaseteli/employer/backend-api/handle-response';
 import useBackendAPI from 'kesaseteli/employer/hooks/useBackendAPI';
 import Company from 'kesaseteli/employer/types/company';
@@ -7,7 +7,7 @@ import { useQuery, UseQueryResult } from 'react-query';
 const useCompanyQuery = (): UseQueryResult<Company, Error> => {
   const { axios } = useBackendAPI();
   return useQuery<Company, Error>('company', () =>
-    handleResponse(axios.get<Company>(backendEndpoint.COMPANY))
+    handleResponse(axios.get<Company>(BackendEndpoint.COMPANY))
   );
 };
 export default useCompanyQuery;
